fix: extract shared User Library path resolver, fix CI
Two CI breakages from PR #115 merge:

1. Format: the multi-line `if` in deploy() wasn't run through Prettier
   before commit. Re-running `npm run format` reformats it.

2. Scripts duplication: 1.66% > 0.7% threshold. The User Library path
   resolver (~25 lines) was duplicated between install-device.ts and
   dev-hot.ts. Extract into scripts/shared/user-library-path.ts; both
   callers import from there. Duplication drops to 0.28%.

install-device.ts previously exited inline on unsupported platforms; it
now treats `null` from the shared resolver as the unsupported case and
exits with the same message.
